switch from `log` to `tracing`
Other Linkerd Rust projects use the [`tracing`] crate for logging,
rather than the `log` crate. This commit updates the failover controller
to use `tracing`.

I tried to make this change as minimal as possible, without changing the
existing logging. In follow-up changes, though, we'll probably want to
change the logging code to use more features of `tracing`, such as
adding spans to the controller's main operations. Also, other
controllers (and the proxy) support a JSON logging mode in addition to
human-readable text, so we may want to add that as well.

[`tracing`]: https://crates.io/crates/tracing