spawn controllers in separate tasks
Currently, the Endpoints and TrafficSplits controllers are run in the
`main` task. This means they cannot run in parallel.

This commit changes the failover controller to spawn the Endpoints and
TrafficSplits controllers in separate `tokio` [tasks]. [Spawning] these
futures as their own tasks allows them to (potentially) run in parallel.

Spawning a task returns a [`JoinHandle`], a future that can be jused to
await the task completing. The `JoinHandle` will complete with an error
instead if the task panics. Since either controller panicking is a fatal
error and the failover controller should panic, I unwrap the `Result`s
returned by these `JoinHandle`s. If both tasks complete normally, the
controller exits gracefully instead.

Additionally, I [instrumented] the futures with [`tracing` spans] prior
to spawning them. This way, any events that are recorded inside the
controllers will be recorded in the context of a span that indicates
_which_ controller the event occurred in. This should make it much
easier to interpret logs from this code, as we can always determine
which controller logged a particular event.

[tasks]: https://docs.rs/tokio/latest/tokio/task/index.html#what-are-tasks
[Spawning]: https://docs.rs/tokio/latest/tokio/task/index.html#spawning
[`JoinHandle`]: https://docs.rs/tokio/latest/tokio/task/struct.JoinHandle.html
[instrumented]: https://docs.rs/tracing/latest/tracing/trait.Instrument.html#method.instrument
[`tracing` spans]: https://docs.rs/tracing/latest/tracing/span/index.html