Find primary svc from annotation (#31)
Fixes #7

We refactor the failover controller to eliminate the assumption that there will only ever be one traffic split watched at any given time.  This means removing most of the fields from the context Data object which previously acted as global state.  Instead, we rely on the kube-rs Store caches.

In order to do this, we need access to both the endpoints and trafficsplit caches whenever either an endpoint or trafficsplit is updated.  Ideally this would be possible by adding a "watches" relationshop to say that the trafficsplit controller watches endpoints. Unfortunately, endpoint updates are not stored in this case and therefore we would not have a populated cache of endpoints when doing trafficsplit reconciliation.

Therefore, we replace the endpoint controller with a reflector which does the following on endpoint updates:

* populates the endpoint cache
* finds any trafficsplits which reference this endpoint, and triggers reconciliation on them

This simplifies the reconciliation logic because the TS reconciler has access to an up-to-date cache of endpoints. 

Signed-off-by: Alex Leong <alex@buoyant.io>

Co-authored-by: Oliver Gould <ver@buoyant.io>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>