Test update behavior (#66)
This change updates the watch-handling code to avoid patching resources
directly. Instead, writes are inserted into a channel and consumed on a
separate task. This was initially done to aid testing--so that we can
consume the channel to validate what patches would be written--but it
has the added benefit of serializing all writes so that concurrent
updates on the endpoints and trafficsplit watches are not acted upon
simultaneously.

These tests highlighted a few issues, which have been fixed:

* A backend could be considered active if it did not have ready
  addresses. We do not want to consider a backend active if it only has
  subsets with `NotReadyAddresses` set.
* We would issue patches even if the traffic split did not actually
  change. We now only issues patches if a backend weight has changed.

This change also adds a _test_ CI workflow to exercises these tests on
pull requests.

Signed-off-by: Oliver Gould <ver@buoyant.io>