Process watch events on a single task (#74)
We currently run the traffic split and endpoints watch on separate tasks
so they can be processed concurrently. But this means that the cache
state can change while processing an update.

For example, as we iterate through a traffic splits backends, it's
possible for the primary service's state to change. While I don't think
this would actively cause bugs given the current state of our code,
reasoning about this dynamic state is difficult.

This change simplifies the behavior by running both watches on a single
task so that events are not processed concurrently.

Signed-off-by: Oliver Gould <ver@buoyant.io>