chore: update `kubert` to 0.9 and `kube` to 0.74
This branch updates `kubert` to v0.9 and `kube` to v0.74. This has to be
done manually as these dependencies must be updated in sync, so they
cannot be updated by Dependabot.

In addition, `kube` deprecated the `ResourceExt::name` method in 0.74,
so I've replaced uses of that method with `ResourceExt::name_any`.

Closes #295
Closes #359