release: Build static binaries  (#432)
Currently, both the CLI and controller builds produce binaries that link
against libc. This necessitates that runtime dependencies be provided by
the operating system (or container base image).

This change updates the build to produce static binaries that are
linkerd against MUSL instead of GNU libc. This change also updates the
build to link against a vendored version of OpenSSL, so that the OpenSSL
version is pinned at build-time.

To do so, this change:

- Updates the devcontainer to v33, which includes a variety of
  build/test utilties.
- Adds a `justfile` that includes recipes for buildling (and
  cross-compiling) CLI binaries
- Moves the `linkerd-failover-cli` binary so that it is built as
  `linkerd-failover` (which is how it should actually be invoked by
  users).
- Removes the local 'docker-build' action, which is copied from the
  Linkerd2 repo.
- Updates CI checks to validate devcontainer images (instead of rust
  toolchains)

See https://github.com/linkerd/dev/pull/13

Signed-off-by: Oliver Gould <ver@buoyant.io>