Replace `curlimages/curl` with `extension-init`
Replaced `curlimages/curl` docker image in the `namespace-metadata` Job
with linkerd's `extension-init` image, to avoid all the OS luggage
included in the former, which generates CVE alerts.

As an added benefit, this extracts the control-planes CNI config to be
able to set the PSA's `pod-security.kubernetes.io/enforce` label in the
`linkerd-failover` namespace.